revert structure change
