+++ globals/helpers/handleApproveDocument.js
