+++ globals/helpers/handleCancelDocument.js
