+++ globals/helpers/handleCreateDocument.js
