+++ globals/helpers/handleCreateOperation.js
