+++ globals/helpers/handleResponse.js
