+++ globals/helpers/handleUpdateDocument.js
