+++ globals/index.js
@@ -1,10 +1,10 @@
-const db = require('./db')
+const mysqlConfig = require('./mysqlConfig')
 const common = require('./common')
 const helpers = require('./helpers')
 const types = require('./types')
 
 module.exports = {
-  db,
+  mysqlConfig,
   helpers,
   types,
   ...common,
